Add a customizable name for the molecule instance
The instance name need to be customizable if we want to run multiple parallel tests.
If multiple tests are launched with the same instance name, there will be collisions.
Thus, allow it to be overrided and feed by something unique (like GitHub github.run_id or Gitlab $CI_JOB_ID)